Implementation of Trace and TraceNode
In this stage, Trace and TraceNode are ready to be generated in the Tracer. Many of the classes is added to support the functionality of them as well.

Due to the amount of code and limited time, we do not do any code refactoring. Most of the code is directly copied from Microbat.

pom.xml is also modified to import eclipse jdt library, which seems to be very important in the old Microbat program.